backend/requirements.txt
fastapi
uvicorn
openai-whisper
spacy
python-multipart
